Move display registration files to a specific folder
