+++ src/display-registration/claim-id-submittor.js
