+++ src/display-registration/claim-id.html
