+++ src/display-registration/display-id-validator.js
