+++ src/display-registration/registration.js
